% Generated by roxygen2 (4.0.1): do not edit by hand
\name{APIurl}
\alias{APIurl}
\title{Get API URL}
\usage{
APIurl(type)
}
\arguments{
\item{type}{a string to paste on the end of the API URI}
}
\value{
the URL as a string
}
\description{
Get the URL for the API
}
\details{
Construct and return the URL for a particular part of the API
}
\author{
Barry Rowlingson
}

