% Generated by roxygen2 (4.0.1): do not edit by hand
\name{position_to_w3w}
\alias{position_to_w3w}
\title{Get what3words for a location}
\usage{
position_to_w3w(xy)
}
\arguments{
\item{xy}{longitude, latitude pair in decimal degrees, in WGS84 coordinates}
}
\value{
a triplet of dotted words
}
\description{
Return the 3words for a location
}
\details{
Use the what3words API to get the 3words for a location
}
\author{
Barry Rowlingson
}

