% Generated by roxygen2 (4.0.1): do not edit by hand
\name{w3w_to_position}
\alias{w3w_to_position}
\title{Get Position from What3Words}
\usage{
w3w_to_position(w)
}
\arguments{
\item{w}{a triplet of dotted words, such as prom.cape.pump}
}
\value{
a data frame of x and y coordinates
}
\description{
Go from 3words or 1word to location
}
\details{
Given a triplet of dotted words or a starred oneword, return the position
}
\author{
Barry Rowlingson
}

